+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Fixed
 - Fixed `BayesStopping` in `stopping` check fitted if score-based
+- Fixed `itr` in `utilities` output shape is input shape
 
 ## Version 1.8.3 (21-05-2025)